Day 5 part 2: now faster, but still ridiculous.
After going through a bunch of different ideas for using concurrency
features, I decided maybe I should run grep on more than one line at a
time, and then I went through a lot of ideas about splitting and merging
streams of maybe-filtered lines before remembering that the normal way to
intersect two `grep`s is to just pipeline them together.  Which is kind
of boring, but effective.

I almost want to try using `mpsc` to get more parallelism than just the
simple pipeline, but maybe there are other things in my life I should be
doing instead....

(N.B. If I actually wanted this to go fast, this is not how I'd do it.)

+++ src/day5.rs
@@ -213,7 +213,7 @@ fn dump() {
 }
 fn checker<S: Scanner + Clone>(s: S) {
     let stdin = stdin();
-    let mut count: usize = 0;
+    let mut count: u64 = 0;
     for line in stdin.lock().lines() {
         let line = line.expect("I/O error reading stdin");
         if nice(s.clone(), &line) {
